+++ src/Entities/AccessTokenEntityInterface.php
@@ -1,0 +1,6 @@
+<?php
+namespace DalPraS\OpenId\Server\Entities;
+
+interface AccessTokenEntityInterface extends 
+    \League\OAuth2\Server\Entities\AccessTokenEntityInterface {
+}