+++ src/Entities/UserEntityInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace DalPraS\OpenId\Server\Entities;
+
+interface UserEntityInterface extends 
+    \League\OAuth2\Server\Entities\UserEntityInterface, 
+    \DalPraS\OpenId\Server\Entities\ClaimSetInterface {
+
+}