+++ src/ResponseTypes/OidcJwtResponse.php
@@ -9,14 +9,14 @@ use Lcobucci\JWT\Builder;
 use Lcobucci\JWT\Signer\Key;
 use Lcobucci\JWT\Signer\Rsa\Sha256;
 use League\OAuth2\Server\Entities\AccessTokenEntityInterface;
-use League\OAuth2\Server\Entities\UserEntityInterface;
 use League\OAuth2\Server\ResponseTypes\BearerTokenResponse;
+use DalPraS\OpenId\Server\Entities\UserEntityInterface;
 
 /**
  * Extends the BearerTokenResponse for adding
  * the param tokenId needed in OpenId.
  */
-class IdTokenJwtResponse extends BearerTokenResponse
+class OidcJwtResponse extends BearerTokenResponse
 {
     /**
      * @var IdentityProviderInterface
@@ -27,22 +27,49 @@ class IdTokenJwtResponse extends BearerTokenResponse
      * @var ClaimExtractor
      */
     protected $claimExtractor;
-
+    
+    /**
+     * @var string
+     */
+    protected $nonce;
+    
     public function __construct(IdentityProviderInterface $identityProvider, ClaimExtractor $claimExtractor) {
         $this->identityProvider = $identityProvider;
         $this->claimExtractor   = $claimExtractor;
     }
-
-    protected function getBuilder(AccessTokenEntityInterface $accessToken, UserEntityInterface $userEntity)
+    
+    /**
+     * @return string
+     */
+    public function getNonce()
     {
-        // Add required id_token claims
-        $builder = (new Builder())
-            ->setAudience($accessToken->getClient()->getIdentifier())
-            ->setIssuer('https://' . $_SERVER['HTTP_HOST'])
-            ->setIssuedAt(time())
-            ->setExpiration($accessToken->getExpiryDateTime()->getTimestamp())
-            ->setSubject($userEntity->getIdentifier());
+        return $this->nonce;
+    }
 
+    /**
+     * @param string $nonce
+     */
+    public function setNonce($nonce)
+    {
+        $this->nonce = $nonce;
+    }
+
+    /**
+     * Get Lcobucci Builder
+     * 
+     * @param AccessTokenEntityInterface $accessToken
+     * @param UserEntityInterface $userEntity
+     * @return \Lcobucci\JWT\Builder
+     */
+    protected function getBuilder(AccessTokenEntityInterface $accessToken, UserEntityInterface $userEntity) {
+        $builder = new Builder();
+        // Add required id_token claims
+        $builder->permittedFor($accessToken->getClient()->getIdentifier())
+                ->issuedBy('https://' . $_SERVER['HTTP_HOST'])
+                ->issuedAt(time())
+                ->expiresAt($accessToken->getExpiryDateTime()->getTimestamp())
+                ->relatedTo($userEntity->getIdentifier())
+        ;
         return $builder;
     }
 
@@ -52,21 +79,17 @@ class IdTokenJwtResponse extends BearerTokenResponse
      */
     protected function getExtraParams(AccessTokenEntityInterface $accessToken)
     {
-        if (false === $this->isOpenIDRequest($accessToken->getScopes())) {
+        if (false === $this->isOidcRequest($accessToken->getScopes())) {
             return [];
         }
 
-        /** @var UserEntityInterface $userEntity */
+        /** @var \League\OAuth2\Server\Entities\UserEntityInterface $userEntity */
         $userEntity = $this->identityProvider->getUserEntityByIdentifier($accessToken->getUserIdentifier());
 
-        if (false === is_a($userEntity, UserEntityInterface::class)) {
-            throw new \RuntimeException('UserEntity must implement UserEntityInterface');
-        } else if (false === is_a($userEntity, ClaimSetInterface::class)) {
-            throw new \RuntimeException('UserEntity must implement ClaimSetInterface');
+        switch (false) {
+            case is_a($userEntity, UserEntityInterface::class):
+                throw new \RuntimeException('UserEntity must implement UserEntityInterface');
         }
-
-        // Add required id_token claims
-        $builder = $this->getBuilder($accessToken, $userEntity);
 
         // Need a claim factory here to reduce the number of claims by provided scope.
         $claims = $this->claimExtractor->extract($accessToken->getScopes(), $userEntity->getClaims());
@@ -76,13 +99,17 @@ class IdTokenJwtResponse extends BearerTokenResponse
             throw new \RuntimeException('UserEntity must set the value of "sub" claim');
         }
 
+        // Add required id_token claims
+        $builder = $this->getBuilder($accessToken, $userEntity);
+
         foreach ($claims as $claimName => $claimValue) {
-            $builder->set($claimName, $claimValue);
+            $builder->withClaim($claimName, $claimValue);
+        }
+        if ( ($nonce = $this->getNonce()) ) {
+            $builder->withClaim('nonce', $nonce);
         }
 
-        $token = $builder
-            ->sign(new Sha256(), new Key($this->privateKey->getKeyPath(), $this->privateKey->getPassPhrase()))
-            ->getToken();
+        $token = $builder->getToken(new Sha256(), new Key($this->privateKey->getKeyPath(), $this->privateKey->getPassPhrase()));
 
         return [
             'id_token' => (string) $token
@@ -95,7 +122,7 @@ class IdTokenJwtResponse extends BearerTokenResponse
      * @param ScopeEntityInterface[] $scopes
      * @return bool
      */
-    private function isOpenIDRequest($scopes) {
+    private function isOidcRequest($scopes) {
         foreach ($scopes as $scope) {
             if ($scope->getIdentifier() === 'openid') {
                 return true;